Jamsine 2.0.0 skeleton
Downloaded Jasmine 2.0.0 and put it under version control.
Downloaded from

https://github.com/pivotal/jasmine/blob/master/dist/jasmine-standalone-2.0.0.zip
The part of the documentation that tells you where to get this stuff is
http://jasmine.github.io/2.0/introduction.html#standalone-distribution

Motivation for this skeleton:  As a developer, it is much easier for me to
clone a repo instead of downloading a zip, so I am making this availible as
a github repo instead of a zip file.