Added actual Jasmine 2.0.0 zip file download link to readme
In order to download a zipfile, you have to set a setting called view raw=true

Why?  Because Github tries to be clever. When you say I want to download a zip file, github generally doesn't believe you. In order for you to tell it that you really really want to downlaod that zip file, you have to click the view raw link, which sets the setting to true and starts the downloaw.