Styling the cookies and ovens with CSS
I copied the css from this fiddle:
http://jsfiddle.net/tararoys/773A7/

1. The next thing to do is to add the css to the Jasmine project.
2. I am creating a folder called css
3. I am adding the jsfiddle css to it and I am linking the specrunner, which now contains my Cookies and Ovens hmtl, to this layout.css file.

Changes I made to the css:
I noticed that the CSS float styling on the fiddle was causing the jasmine tests to overlap on top of the testing suite, so for the sake of simplicity I got rid of the float styling.
I also put the html portion inside it's own bakery div with a bakery class
I edited the selectors of in layout.css so that, theoretically, they only apply to descendant elements of the bakery class. I'm kind of fuzzy on how they work, so I used this tutorial for how to do descendants.