Don't try to put your html webpage in your jasmine specrunner
The point of Jasmine is to be able to test your javascript in isolation.
Cramming the code for my cookies into the specrunner so that I could
see the page and then the tests for that page below it broke how
jasmine thinks things should work.

Jasmine is pretty much concerned with testing only the things taht are
in your /src file.

The /src file is the standard place where most webpages store their javascript.

So this project really has two parts:

the jasmine testing framework, which consists right now of the
  1. Jasmine 2.0.0 in the lib folder
  2. the spec folder
  3. All the javascripts in the src folder
  4. And the specrunner.

  The specrunner's job is to look in the spec runner for tests, and
  run those tests on the javascripts in the test folder.

The second part of this project is your actual webpage.  It consists of
  1. the css folder
  2. the index.html page
  3. The javascripts in src
  4. (eventually you could store other code in the lib folder, like, for example, jquery. )

This collection of files and foders is your actual website.

So...jasmine's job is to make sure the javascripts in the src file actually work.

Jasmine is like a teacher who grades your paper based on a rubric to make sure you are doing good work.