Take out default example code and include Cookie example code.
It makes more sense for this Cookies and Ovens setup skeleton to show
how to link a Cookie spec and and Cookies.js file using the specrunner
than it does to have the specrunner, the spec folder, and the source
folder to continue to have the example tests.

This commit shows how to link up a testing suite for a new object.